Use INT type instead of NUMERIC as the latter is a decimal in PostgreSQL

+++ aquadoggo/migrations/20220509090252_create-operations.sql
@@ -15,7 +15,7 @@ CREATE TABLE IF NOT EXISTS operation_fields_v1 (
     name                    TEXT             NOT NULL,
     field_type              TEXT             NOT NULL,
     value                   TEXT             NULL,
-    list_index              NUMERIC          NOT NULL,
+    list_index              INT              NOT NULL,
     FOREIGN KEY(operation_id) REFERENCES operations_v1(operation_id)
 );
